+++ src/main/java/ledger/user_interface/utils/AAmountStringConverter.java
@@ -5,18 +5,18 @@ import javafx.util.StringConverter;
 /**
  * Created by Tayler How on 2/5/2017.
  */
-public abstract class AAmountStringConverter extends StringConverter<Integer> {
+public abstract class AAmountStringConverter extends StringConverter<Long> {
 
-    public Integer fromString(String amountString) {
+    public Long fromString(String amountString) {
         if (InputSanitization.isInvalidAmount(amountString)) {
             // invalid amount
             return null;
         }
 
         double amountToSetDecimal = Double.parseDouble(amountString);
-        int amount = (int) Math.round(amountToSetDecimal * 100);
+        long amount = (long) Math.round(amountToSetDecimal * 100);
         return amount;
     }
 
-    public abstract String toString(Integer amount);
+    public abstract String toString(Long amount);
 }
